Export berth applications to Excel
Refs VEN-1450
