+++ exports/views.py
@@ -8,11 +8,13 @@ from rest_framework.authentication import SessionAuthentication
 from rest_framework.permissions import DjangoModelPermissions
 from rest_framework.views import APIView
 
+from applications.models import BerthApplication
+from applications.schema import BerthApplicationNode
 from berth_reservations.oidc import BerthApiTokenAuthentication
 from customers.models import CustomerProfile
 from customers.schema import ProfileNode
 from exports.utils import from_global_ids
-from exports.xlsx_writer import BaseExportXlsxWriter, CustomerXlsx
+from exports.xlsx_writer import BaseExportXlsxWriter, BerthApplicationXlsx, CustomerXlsx
 
 
 class UserHasViewPermission(DjangoModelPermissions):
@@ -90,3 +92,13 @@ class CustomerExportView(BaseExportView):
             raise serializers.ValidationError("profile_token required.")
 
         return kwargs
+
+
+class BerthApplicationExportView(BaseExportView):
+    model = BerthApplication
+    exporter_class = BerthApplicationXlsx
+
+    def get_queryset(self, ids: Iterable = None):
+        if ids:
+            ids = from_global_ids(ids, BerthApplicationNode)
+        return super().get_queryset(ids=ids)
